// src/yeying/api/bulletin/bulletin.ts
export interface SolutionMetadata {
  publisher: string;
  language: LanguageCodeEnum;
  uid: string;
  name: string;
  description: string;
  createdAt: string;
  cards: SolutionCard[];
  signature: string;
}

export interface SolutionCard {
  name: string;
  price: string;
  variables: string;
}
// ...
export const SolutionMetadata: MessageFns<SolutionMetadata> = {
  encode(message: SolutionMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.publisher !== "") {
      writer.uint32(10).string(message.publisher);
    }
    if (message.language !== 0) {
      writer.uint32(16).int32(message.language);
    }
    if (message.uid !== "") {
      writer.uint32(26).string(message.uid);
    }
    if (message.name !== "") {
      writer.uint32(34).string(message.name);
    }
    if (message.description !== "") {
      writer.uint32(42).string(message.description);
    }
    if (message.createdAt !== "") {
      writer.uint32(50).string(message.createdAt);
    }
    for (const v of message.cards) {
      SolutionCard.encode(v!, writer.uint32(58).fork()).join();
    }
    if (message.signature !== "") {
      writer.uint32(66).string(message.signature);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SolutionMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSolutionMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.publisher = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.language = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.uid = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.cards.push(SolutionCard.decode(reader, reader.uint32()));
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.signature = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SolutionMetadata {
    return {
      publisher: isSet(object.publisher) ? globalThis.String(object.publisher) : "",
      language: isSet(object.language) ? languageCodeEnumFromJSON(object.language) : 0,
      uid: isSet(object.uid) ? globalThis.String(object.uid) : "",
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      description: isSet(object.description) ? globalThis.String(object.description) : "",
      createdAt: isSet(object.createdAt) ? globalThis.String(object.createdAt) : "",
      cards: globalThis.Array.isArray(object?.cards) ? object.cards.map((e: any) => SolutionCard.fromJSON(e)) : [],
      signature: isSet(object.signature) ? globalThis.String(object.signature) : "",
    };
  },

  toJSON(message: SolutionMetadata): unknown {
    const obj: any = {};
    if (message.publisher !== "") {
      obj.publisher = message.publisher;
    }
    if (message.language !== 0) {
      obj.language = languageCodeEnumToJSON(message.language);
    }
    if (message.uid !== "") {
      obj.uid = message.uid;
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.description !== "") {
      obj.description = message.description;
    }
    if (message.createdAt !== "") {
      obj.createdAt = message.createdAt;
    }
    if (message.cards?.length) {
      obj.cards = message.cards.map((e) => SolutionCard.toJSON(e));
    }
    if (message.signature !== "") {
      obj.signature = message.signature;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SolutionMetadata>, I>>(base?: I): SolutionMetadata {
    return SolutionMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SolutionMetadata>, I>>(object: I): SolutionMetadata {
    const message = createBaseSolutionMetadata();
    message.publisher = object.publisher ?? "";
    message.language = object.language ?? 0;
    message.uid = object.uid ?? "";
    message.name = object.name ?? "";
    message.description = object.description ?? "";
    message.createdAt = object.createdAt ?? "";
    message.cards = object.cards?.map((e) => SolutionCard.fromPartial(e)) || [];
    message.signature = object.signature ?? "";
    return message;
  },
};

function createBaseSolutionCard(): SolutionCard {
  return { name: "", price: "", variables: "" };
}

export const SolutionCard: MessageFns<SolutionCard> = {
  encode(message: SolutionCard, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.price !== "") {
      writer.uint32(18).string(message.price);
    }
    if (message.variables !== "") {
      writer.uint32(26).string(message.variables);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SolutionCard {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSolutionCard();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.price = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.variables = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SolutionCard {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      price: isSet(object.price) ? globalThis.String(object.price) : "",
      variables: isSet(object.variables) ? globalThis.String(object.variables) : "",
    };
  },

  toJSON(message: SolutionCard): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.price !== "") {
      obj.price = message.price;
    }
    if (message.variables !== "") {
      obj.variables = message.variables;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SolutionCard>, I>>(base?: I): SolutionCard {
    return SolutionCard.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SolutionCard>, I>>(object: I): SolutionCard {
    const message = createBaseSolutionCard();
    message.name = object.name ?? "";
    message.price = object.price ?? "";
    message.variables = object.variables ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}